Separated a monolithic jello.js file into separate smaller files.
