+++ lib/jello.files.js
@@ -1,0 +1,9 @@
+Jello.Files = function () {
+  var _private = (function () {
+    return{
+
+    };
+  })();
+  
+  throw ("Not implemented");
+};
