+++ lib/jello.list.js
@@ -1,0 +1,9 @@
+Jello.List = function () {
+  var _private = (function () {
+    return{
+
+    };
+  })();
+
+  throw ("Not implemented");
+};
