+++ lib/jello.web.js
@@ -1,0 +1,9 @@
+Jello.Web = function () {
+  var _private = (function () {
+    return{
+
+    };
+  })();
+
+  throw ("Not implemented");
+};
