+++ package.json
@@ -27,6 +27,7 @@
   "homepage": "https://github.com/omkarkhair/sp-jello#readme",
   "devDependencies": {
     "gulp": "^3.9.1",
+    "gulp-concat": "^2.6.1",
     "gulp-minify": "0.0.14"
   },
   "dependencies": {
